+++ README.md
@@ -1,94 +1,97 @@
-# IBM Enterprise Module
+# IBM Enterprise Account module
 
 This is a collection of sub modules and which will take tree based enterprise inputs and create child accounts and account-groups in an enterprise
 * [dynamic_values](submodules/dynamic_values)
 * [enterprise_hierarchy](submodules/enterprise_hierarchy)
 
-
+<!-- Below content is automatically populated via pre-commit hook -->
+<!-- BEGIN OVERVIEW HOOK -->
 ## Overview
 * [terraform-ibm-enterprise](#terraform-ibm-enterprise)
 * [Examples](./examples)
+    * [Advanced example](./examples/advanced)
     * [Basic example](./examples/basic)
-    * [Complete example](./examples/complete)
 * [Contributing](#contributing)
+<!-- END OVERVIEW HOOK -->
 
-## Compatibility
+<!-- This heading should always match the name of the root level module (aka the repo name) -->
+## terraform-ibm-enterprise
 
-This module is meant for use with Terraform 0.13 (and higher).
+### Usage
 
-## Usage
-
-Full examples are in the [examples](./examples/) folder, but basic usage is as follows for creation of enterprise childrens is
+Full examples are in the [examples](./examples/) folder, but basic usage is as follows for creation of enterprise children is
 
 ```hcl
 provider "ibm" {
-ibmcloud_api_key = var.ibmcloud_api_key # pragma: allowlist secret
+  ibmcloud_api_key = "XXXXXXXX" # pragma: allowlist secret
 }
 
 data "ibm_enterprises" "enterprise" {
-  name = var.enterprise_name
+  name = "my-enterprise-account"
 }
 
 module "enterprise" {
-  source                       = "terraform-ibm-modules/terraform-ibm-enterprise"
-  enterprise_crn       = data.ibm_enterprises.enterprise.enterprises[0].crn
+  source                            = "terraform-ibm-modules/terraform-ibm-enterprise"
+  enterprise_crn                    = data.ibm_enterprises.enterprise.enterprises[0].crn
   enterprise_primary_contact_iam_id = data.ibm_enterprises.enterprise.enterprises[0].primary_contact_iam_id
-  enterprise_json_input = var.enterprise_json_input
-}
-
+  enterprise_json_input             = {
+                                        account_groups = [{
+                                          name = "depth_0_account_group1"
+                                        }]
+                                        accounts = [{
+                                          name = "depth_0_account" },
+                                        ]}
+                                      }
 ```
 
-## Requirements
+### Required IAM access policies
 
-### Terraform plugins
+- Account Management
+  - **Enterprise** service
+      - `Administrator` platform access
 
-- [Terraform](https://www.terraform.io/downloads.html) 0.13 (or later)
-- [terraform-provider-ibm](https://github.com/IBM-Cloud/terraform-provider-ibm)
+<!-- Below content is automatically populated via pre-commit hook -->
+<!-- BEGINNING OF PRE-COMMIT-TERRAFORM DOCS HOOK -->
+### Requirements
 
-## Install
+| Name | Version |
+|------|---------|
+| <a name="requirement_terraform"></a> [terraform](#requirement\_terraform) | >= 1.3.0, <1.6.0 |
+| <a name="requirement_ibm"></a> [ibm](#requirement\_ibm) | >= 1.54.0, < 2.0.0 |
 
-### Terraform
+### Modules
 
-Be sure you have the correct Terraform version (0.13), you can choose the binary here:
-- https://releases.hashicorp.com/terraform/
+| Name | Source | Version |
+|------|--------|---------|
+| <a name="module_dynamic_values"></a> [dynamic\_values](#module\_dynamic\_values) | ./submodules/dynamic_values | n/a |
+| <a name="module_enterprise_hierarchy_depth_0"></a> [enterprise\_hierarchy\_depth\_0](#module\_enterprise\_hierarchy\_depth\_0) | ./submodules/enterprise_hierarchy | n/a |
+| <a name="module_enterprise_hierarchy_depth_1"></a> [enterprise\_hierarchy\_depth\_1](#module\_enterprise\_hierarchy\_depth\_1) | ./submodules/enterprise_hierarchy | n/a |
+| <a name="module_enterprise_hierarchy_depth_2"></a> [enterprise\_hierarchy\_depth\_2](#module\_enterprise\_hierarchy\_depth\_2) | ./submodules/enterprise_hierarchy | n/a |
 
-### Terraform plugins
+### Resources
 
-Be sure you have the compiled plugins on $HOME/.terraform.d/plugins/
+No resources.
 
-- [terraform-provider-ibm](https://github.com/IBM-Cloud/terraform-provider-ibm)
+### Inputs
 
-### Pre-commit hooks
+| Name | Description | Type | Default | Required |
+|------|-------------|------|---------|:--------:|
+| <a name="input_enterprise_crn"></a> [enterprise\_crn](#input\_enterprise\_crn) | The CRN of the parent Enterprise account to use. | `string` | n/a | yes |
+| <a name="input_enterprise_json_input"></a> [enterprise\_json\_input](#input\_enterprise\_json\_input) | List of account groups and account names | <pre>map(list(object({<br>    name         = string<br>    owner_iam_id = optional(string, "root")<br>    accounts     = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))<br>    account_groups = optional(list(object(<br>      { name     = string,<br>        accounts = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))<br>        account_groups = optional(list(object({<br>          name     = string,<br>          accounts = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))<br>          account_groups = optional(list(object(<br>          { name = string })))<br>        })))<br>    })))<br>    }))<br>  )</pre> | n/a | yes |
+| <a name="input_enterprise_primary_contact_iam_id"></a> [enterprise\_primary\_contact\_iam\_id](#input\_enterprise\_primary\_contact\_iam\_id) | The IAM id of the parent Enterprise account owner. | `string` | n/a | yes |
 
-Run the following command to execute the pre-commit hooks defined in .pre-commit-config.yaml file
-```
-pre-commit run -a
-```
-You can install pre-coomit tool using
+### Outputs
 
-```
-pip install pre-commit
-```
-or
-```
-pip3 install pre-commit
-```
-## How to input variable values through a file
+| Name | Description |
+|------|-------------|
+| <a name="output_enterprise_hierarchy_depth_0"></a> [enterprise\_hierarchy\_depth\_0](#output\_enterprise\_hierarchy\_depth\_0) | Enterprise input for depth 0 |
+| <a name="output_enterprise_hierarchy_depth_1"></a> [enterprise\_hierarchy\_depth\_1](#output\_enterprise\_hierarchy\_depth\_1) | Enterprise input for depth 1 |
+| <a name="output_enterprise_hierarchy_depth_2"></a> [enterprise\_hierarchy\_depth\_2](#output\_enterprise\_hierarchy\_depth\_2) | Enterprise input for depth 2 |
+<!-- END OF PRE-COMMIT-TERRAFORM DOCS HOOK -->
 
-To review the plan for the configuration defined (no resources actually provisioned)
-```
-terraform plan -var-file=./input.tfvars
-```
-To execute and start building the configuration defined in the plan (provisions resources)
-```
-terraform apply -var-file=./input.tfvars
-```
+<!-- Leave this section as is so that your module has a link to local development environment set up steps for contributors to follow -->
+## Contributing
 
-To destroy the VPC and all related resources
-```
-terraform destroy -var-file=./input.tfvars
-```
+You can report issues and request features for this module in GitHub issues in the module repo. See [Report an issue or request a feature](https://github.com/terraform-ibm-modules/.github/blob/main/.github/SUPPORT.md).
 
-## Note
-
-All optional parameters, by default, will be set to `null` in respective example's varaible.tf file. You can also override these optional parameters.
+To set up your local development environment, see [Local development setup](https://terraform-ibm-modules.github.io/documentation/#/local-dev-setup) in the project documentation.
