+++ examples/advanced/README.md
@@ -1,0 +1,3 @@
+# Advanced example
+
+This example will create a hierarchy of enterprise child accounts and account groups.
