+++ examples/advanced/outputs.tf
