+++ examples/advanced/variables.tf
