--- examples/complete/README.md
@@ -1,3 +1,0 @@
-# Complete example
-
-This example will create a hierarchy of enterprise child accounts and account groups
