--- examples/complete/variables.tf
@@ -1,38 +1,0 @@
-variable "enterprise_name" {
-  description = "name of the enterprise"
-  type        = string
-  default     = "DAF Enterprise"
-}
-
-# for now added support for upto 3 level depth of enterprise hierarchy
-variable "enterprise_json_input" {
-  description = "List of enterprise child account groups and account"
-  type = map(list(object({
-    name         = string
-    owner_iam_id = optional(string, "root")
-    accounts     = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))
-    account_groups = optional(list(object(
-      { name     = string,
-        accounts = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))
-        account_groups = optional(list(object({
-          name     = string,
-          accounts = optional(list(object({ name = string, owner_iam_id = optional(string, "root") })))
-          account_groups = optional(list(object(
-          { name = string })))
-        })))
-    })))
-    }))
-  )
-  default = {
-    account_groups = [{
-      name = "depth_0_account_group1"
-    }]
-    accounts = [{ name = "depth_0_account" },
-  ] }
-}
-
-variable "ibmcloud_api_key" {
-  type        = string
-  description = "The IBM Cloud API key for generating token this account authenticates to"
-  sensitive   = true
-}
