+++ submodules/enterprise_hierarchy/README.md
@@ -1,3 +1,3 @@
-# enterprise hierarchy
+# Enterprise hierarchy
 
-This submodule create enterprise child accounts and account groups at multipled depth of interprise hierarchy
+This submodule create enterprise child accounts and account groups at multiple depths of an enterprise hierarchy
