+++ tests/README.md
@@ -1,8 +1,5 @@
-<!-- BEGIN TESTS HOOK -->
-
 # Tests
 
 For information about how to create and run tests, see [Validation tests](https://terraform-ibm-modules.github.io/documentation/#/tests) in the project documentation.
 
 <!-- Add any more steps that are specific to testing this module and that are not in the docs. -->
-<!-- END TESTS HOOK -->
