+++ tests/pr_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
 )
 
-// Use existing resource group
-const completeExampleDir = "examples/complete"
+const basicExampleDir = "examples/basic"
+const advancedExampleDir = "examples/advanced"
 
 func setupOptions(t *testing.T, prefix string, dir string) *testhelper.TestOptions {
 	options := testhelper.TestOptionsDefault(&testhelper.TestOptions{
@@ -17,15 +17,44 @@ func setupOptions(t *testing.T, prefix string, dir string) *testhelper.TestOptio
 		TerraformDir: dir,
 		Prefix:       prefix,
 	})
+
+	options.TerraformVars = map[string]interface{}{
+		"prefix":          options.Prefix,
+		"enterprise_name": "DAF enterprise",
+	}
+
 	return options
+}
+
+func TestRunBasicExample(t *testing.T) {
+	t.Parallel()
+
+	options := setupOptions(t, "enterprise-basic", basicExampleDir)
+
+	output, err := options.RunTestConsistency()
+	assert.Nil(t, err, "This should not have errored")
+	assert.NotNil(t, output, "Expected some output")
 }
 
 func TestRunCompleteExample(t *testing.T) {
 	t.Parallel()
 
-	options := setupOptions(t, "enterprise-cust", completeExampleDir)
+	options := setupOptions(t, "enterprise-com", advancedExampleDir)
 
 	output, err := options.RunTestConsistency()
 	assert.Nil(t, err, "This should not have errored")
 	assert.NotNil(t, output, "Expected some output")
+}
+
+func TestRunUpgradeExample(t *testing.T) {
+	t.Parallel()
+
+	// TODO: Replace with advanced example when issue with destroy account groups is resolved
+	options := setupOptions(t, "enterprise-upg", basicExampleDir)
+
+	output, err := options.RunTestUpgrade()
+	if !options.UpgradeTestSkipped {
+		assert.Nil(t, err, "This should not have errored")
+		assert.NotNil(t, output, "Expected some output")
+	}
 }
